+++ BBGE/ActionMapper.cpp
@@ -102,7 +102,7 @@ void ActionMapper::addAction(Event *event, int k, int state)
 
 Event* ActionMapper::addCreatedEvent(Event *event)
 {
-	for (int i = 0; i < createdEvents.size(); i++)
+	for (size_t i = 0; i < createdEvents.size(); i++)
 	{
 		if (createdEvents[i] == event)
 			return event;
@@ -113,7 +113,7 @@ Event* ActionMapper::addCreatedEvent(Event *event)
 
 void ActionMapper::clearCreatedEvents()
 {
-	for (int i = 0; i < createdEvents.size(); i++)
+	for (size_t i = 0; i < createdEvents.size(); i++)
 	{
 		delete createdEvents[i];
 	}
@@ -326,7 +326,7 @@ void ActionMapper::removeAllActions()
 	{
 		deleteList.push_back(i->id);
 	}
-	for (int c = 0; c < deleteList.size(); c++)
+	for (size_t c = 0; c < deleteList.size(); c++)
 	{
 		removeAction (deleteList[c]);
 	}
